Allow multiple projects in one set of prompts
* Helps if your change affects multiple projects without being
too inconvenient if not.
* Add an example for projects.
* Add a link to npm package in nodejs install docs
