+++ examples/README.md
@@ -9,3 +9,8 @@ Uses the default configuration generated from `changie init`.
 Uses the [this configuration](./custom_prompts_config.yaml).
 
 ![custom_prompts](./custom_prompts.gif)
+
+## Projects
+Uses the [this configuration](./projects_config.yaml).
+
+![projects](./projects.gif)
