+++ examples/projects_config.yaml
@@ -1,0 +1,38 @@
+changesDir: .changes
+unreleasedDir: unreleased
+headerPath: header.tpl.md
+changelogPath: CHANGELOG.md
+versionExt: md
+versionFormat: '## {{.Version}} on {{.Time.Format "2006-01-02"}}'
+kindFormat: '### {{.Kind}}'
+changeFormat: '* {{.Body}}'
+kinds:
+  - label: Added
+    auto: minor
+  - label: Changed
+    auto: major
+  - label: Deprecated
+    auto: minor
+  - label: Removed
+    auto: major
+  - label: Fixed
+    auto: patch
+  - label: Security
+    auto: patch
+newlines:
+  afterChangelogHeader: 1
+  afterKind: 1
+  afterChangelogVersion: 1
+  beforeKind: 1
+  endOfVersion: 1
+projects:
+  - label: User Feeds
+    key: user_feeds
+    changelog: users/feeds/CHANGELOG.md
+  - label: User Management
+    key: user_management
+    changelog: users/management/CHANGELOG.md
+  - label: Landing Page
+    key: landing
+    changelog: landing/CHANGELOG.md
+projectsVersionSeparator: '/'
